refactor: split repl and turn execution
